Update some dependencies.
Biggest change here is Indicatif, which makes ProgressStyle::template fallible, and removes the explicit join on MultiBar.

It does seem to have fixed the render bug when the bars finish too fast, but I'll keep the clear-and-rerender, as the progress bars render on stderr, while the rerender is on stdout, and I like that behaviour for piping.